+++ CHANGELOG.md
@@ -7,9 +7,16 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 
 ## Unreleased
+
 ### Added
-- `BEAM.Memory` plugin: Monitors memory usage of the VM.
-- `BEAM.SystemInfo` plugin: Monitors system counters of the VM.
+
+- `CrowPlugins.BEAM.ContextSwitches` plugin: Counts total context switches in
+  the VM.
+- `CrowPlugins.BEAM.GarbageCollections` plugin: Counts total garbage collections
+  in the VM.
+- `CrowPlugins.BEAM.IO` plugin: Monitors bytes sent and received through ports.
+- `CrowPlugins.BEAM.Memory` plugin: Monitors memory usage of the VM.
+- `CrowPlugins.BEAM.SystemInfo` plugin: Monitors system counters of the VM.
 
 
 <!-- vim: set textwidth=80 sw=2 ts=2: -->